Clean up task and plugin files
