+++ plugin-build/gradle.properties
@@ -5,4 +5,4 @@ DISPLAY_NAME=Android/Gradle module dependency graphs
 DESCRIPTION=A plugin to produce Github/mermaid compatible dependency graphs for your projects.
 WEBSITE=https://github.com/adityabhaskar/Project-Dependency-Graph
 VCS_URL=https://github.com/adityabhaskar/Project-Dependency-Graph
-IMPLEMENTATION_CLASS=net.c306.dependencygraph.plugin.TemplatePlugin
+IMPLEMENTATION_CLASS=net.c306.dependencygraph.plugin.DependencyGraphPlugin
